requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.4
black==25.11.0
boto3==1.42.39
botocore==1.42.39
celery==5.6.2
cffi==2.0.0
channels==4.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.10.7
cryptography==46.0.4
daphne==4.2.1
Django==4.2.27
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
exceptiongroup==1.3.1
factory_boy==3.3.3
Faker==37.12.0
flake8==7.3.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.1
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.1.0
isort==6.1.0
jmespath==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytokens==0.4.1
PyYAML==6.0.3
redis==7.0.1
referencing==0.36.2
rpds-py==0.27.1
s3transfer==0.16.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tomli==2.4.0
Twisted==25.5.0
txaio==23.6.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20
vine==5.1.0
wcwidth==0.5.3
whitenoise==6.11.0
zipp==3.23.0
zope.interface==8.0.1
